% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.im.SpatRaster.R
\name{as.im.SpatRaster}
\alias{as.im.SpatRaster}
\title{Convert SpatRaster to Pixel Image}
\usage{
\method{as.im}{SpatRaster}(X, ...)
}
\arguments{
\item{X}{a SpatRaster}

\item{...}{additional arguments (unused)}
}
\value{
A pixel image
}
\description{
Converts a SpatRaster to a pixel image
}
